integrate group page into video-chat-room
