+++ src/components/calendar/studentCalendar.tsx
@@ -156,7 +156,7 @@ function Header({ appointmentData, ...restProps }: AppointmentTooltip.HeaderProp
             width={30}
             onClick={() => {
               if (!appointmentData || typeof appointmentData.id !== "string") return;
-              router.push(`/myCourses/classRoom/studentRoom/${appointmentData.id}`);
+              router.push(`/myCourses/classRoom/videoChatRoom/${appointmentData.id}`);
             }}
           />
         </ButtonWrapper>
